requirements.txt
numpy==1.24.3
opencv-python==4.8.0.74
streamlit==1.27.0
Pillow==9.5.0
scikit-image==0.21.0
torch==2.0.1
torchvision==0.15.2 